Species Distributions

A Species Distribution is a tibble::tibble containing species abundances or probabilities.

species_distribution(x, names = NULL, weights = NULL, check_arguments = TRUE)

as_species_distribution(x, ...)

# S3 method for numeric
as_species_distribution(x, ..., check_arguments = TRUE)

# S3 method for matrix
as_species_distribution(x, ..., check_arguments = TRUE)

# S3 method for data.frame
as_species_distribution(x, ..., check_arguments = TRUE)

is_species_distribution(x)

probabilities(
  x,
  estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  jack_max = 10,
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  q = 0,
  check_arguments = TRUE
)

as_probabilities(x, ...)

# S3 method for numeric
as_probabilities(x, ..., check_arguments = TRUE)

# S3 method for matrix
as_probabilities(x, names = NULL, weights = NULL, ..., check_arguments = TRUE)

is_probabilities(x)

abundances(
  x,
  round = TRUE,
  names = NULL,
  weights = NULL,
  check_arguments = TRUE
)

as_abundances(x, ...)

# S3 method for numeric
as_abundances(x, round = TRUE, ..., check_arguments = TRUE)

# S3 method for integer
as_abundances(x, ..., check_arguments = TRUE)

# S3 method for matrix
as_abundances(
  x,
  round = TRUE,
  names = NULL,
  weights = NULL,
  ...,
  check_arguments = TRUE
)

is_abundances(x)

Arguments

x

An object.

names

The names of the species distributions.

weights

The weights of the sites of the species distributions.

check_arguments

If TRUE, the function arguments are verified. Should be set to FALSE to save time when the arguments have been checked elsewhere.

...

Additional arguments to be passed to plot. Unused elsewhere.

estimator

One of the estimators of a probability distribution: "naive" (the default value), or "Chao2013", "Chao2015", "ChaoShen" to estimate the probabilities of the observed species in the asymptotic distribution.

unveiling

One of the possible unveiling methods to estimate the probabilities of the unobserved species: "none" (default, no species is added), "uniform" (all unobserved species have the same probability) or "geometric" (the unobserved species distribution is geometric).

richness_estimator

An estimator of richness to evaluate the total number of species. "jackknife" is the default value. An alternative is "rarefy" to estimate the number of species such that the entropy of the asymptotic distribution rarefied to the observed sample size equals the actual entropy of the data.

jack_max

The highest jackknife order allowed. Default is 10. Allowed values are between 1 and 10.

coverage_estimator

An estimator of sample coverage used by coverage: "ZhangHuang" (the default value), "Chao", "Turing" or "Good".

q

The order of diversity. Default is 0 for richness. Used only to estimate asymptotic probability distributions when argument richness_estimator is "rarefy". Then, the number of unobserved species is fitted so that the entropy of order q of the asymptotic probability distribution at the observed sample size equals the actual entropy of the data.

round

If TRUE, the values of x are converted to integers.

Details

species_distribution objects include abundances and probabilities objects.

as_abundances() formats the numeric or integer x so that appropriate versions of community functions (generic methods such as Diversity) are applied. Abundance values are rounded (by default) to the nearest integer.

as_probabilities() normalizes the vector x so that it sums to 1. If the estimator is not "naive", the observed abundance distribution is used to estimate the actual species distribution. The list of species will be changed: zero-abundance species will be cleared, and some unobserved species will be added. First, observed species probabilities are estimated following Chao and Shen (2003) , i.e. input probabilities are multiplied by the sample coverage, or according to more sophisticated models: Chao et al. (2013) , a single-parameter model, or Chao and Jost (2015) , a two-parameter model. The total probability of observed species equals the sample coverage. Then, the distribution of unobserved species can be unveiled: their number is estimated according to the richness_estimator. The coverage deficit (1 minus the sample coverage) is shared by the unobserved species equally (unveiling = "uniform", (Chao et al. 2013) ) or according to a geometric distribution (unveiling = "geometric", (Chao and Jost 2015) ).

TODO: These functions can be applied to data frames to calculate the joint diversity (Gregorius 2010) .

species_distribution objects objects can be plotted by plot and autoplot().

References

Chao A, Jost L (2015). “Estimating Diversity and Entropy Profiles via Discovery Rates of New Species.” Methods in Ecology and Evolution, 6(8), 873--882. doi:10.1111/2041-210X.12349 .

Chao A, Shen T (2003). “Nonparametric Estimation of Shannon's Index of Diversity When There Are Unseen Species in Sample.” Environmental and Ecological Statistics, 10(4), 429--443. doi:10.1023/A:1026096204727 .

Chao A, Wang Y, Jost L (2013). “Entropy and the Species Accumulation Curve: A Novel Entropy Estimator via Discovery Rates of New Species.” Methods in Ecology and Evolution, 4(11), 1091--1100. doi:10.1111/2041-210x.12108 .

Gregorius H (2010). “Linking Diversity and Differentiation.” Diversity, 2(3), 370--394. doi:10.3390/d2030370 .

Examples

# Paracou data
is_species_distribution(paracou_6_abd)
#> [1] TRUE
# Whittaker plot fitted by a log-normal distribution TODO
# autoplot(asm_paracou_6, distribution = "lognormal")