Abundance Frequency Count of a Community

Count the number of species observed the same number of times.

Usage

abd_freq_count(
  abd,
  level = NULL,
  probability_estimator = c("naive", "Chao2013", "Chao2015", "ChaoShen"),
  unveiling = c("none", "uniform", "geometric"),
  richness_estimator = c("jackknife", "iChao1", "Chao1", "rarefy", "naive"),
  coverage_estimator = c("ZhangHuang", "Chao", "Turing", "Good"),
  check_arguments = TRUE
)

Arguments

abd

A numeric vector containing species abundances.

level

the level of interpolation or extrapolation. It may be a sample size (an integer) or a sample coverage (a number between 0 and 1). If not NULL, the asymptotic estimator is ignored.

probability_estimator

a string containing one of the possible estimators of the probability distribution (see probabilities). Used only for extrapolation.

unveiling

a string containing one of the possible unveiling methods to estimate the probabilities of the unobserved species (see probabilities). Used only for extrapolation.

richness_estimator

A string containing an estimator recognized by div_richness to evaluate the total number of species in probabilities. Used only for extrapolation.

coverage_estimator

an estimator of sample coverage used by coverage.

check_arguments

if TRUE, the function arguments are verified. Should be set to FALSE to save time when the arguments have been checked elsewhere.

Value

A two-column tibble. The first column contains the number of observations, the second one the number of species observed this number of times.

Details

The Abundance Frequency Count Chao2015divent is the number of species observed each number of times. It is a way to summarize the species distribution.

It can be estimated at a specified level of interpolation or extrapolation. Extrapolation relies on the estimation of the estimation of the asymptotic distribution of the community by probabilities and eq. (5) of Chao2014divent.

References

Examples

abd_freq_count(as.numeric(paracou_6_abd[1, ]))
#> # A tibble: 25 × 2
#>    abundance number_of_species
#>        <int>         <int[1d]>
#>  1         1                84
#>  2         2                35
#>  3         3                19
#>  4         4                14
#>  5         5                 5
#>  6         6                 4
#>  7         7                 3
#>  8         8                 3
#>  9        10                 2
#> 10        11                 1
#> # ℹ 15 more rows