check_divent_args

Checks the arguments of a function of the package divent

check_divent_args(
  abundances = NULL,
  as_numeric = NULL,
  check_arguments = NULL,
  estimator = NULL,
  jack_alpha = NULL,
  jack_max = NULL,
  level = NULL,
  n_simulations = NULL,
  probability_estimator = NULL,
  q = NULL,
  richness_estimator = NULL,
  sample_coverage = NULL,
  show_progress = NULL,
  unveiling = NULL
)

Arguments

abundances

An object of class abundances.

as_numeric

If TRUE, a number or a numeric vector is returned rather than a tibble.

check_arguments

If TRUE, the function arguments are verified. Should be set to FALSE to save time when the arguments have been checked elsewhere.

estimator

An estimator of asymptotic entropy, diversity or richness.

jack_alpha

The risk level, 5% by default, used to optimize the jackknife order.

jack_max

The highest jackknife order allowed. Default is 10.

level

The level of interpolation or extrapolation. It may be a sample size (an integer) or a sample coverage (a number between 0 and 1). If not NULL, the asymptotic estimator is ignored.

n_simulations

The number of simulations used to estimate the confidence envelope.

probability_estimator

A string containing one of the possible estimators of the probability distribution (see probabilities). Used only for extrapolation.

q

The order of diversity.

richness_estimator

An estimator of richness to evaluate the total number of species, see div_richness. Used for interpolation and extrapolation.

sample_coverage

The sample coverage of x calculated elsewhere. Used to calculate the gamma diversity of meta-communities, see details.

show_progress

If TRUE, a progress bar is shown during long computations.

unveiling

A string containing one of the possible unveiling methods to estimate the probabilities of the unobserved species (see probabilities). Used only for extrapolation.

alpha

The risk level, 5% by default.

coverage_estimator

An estimator of sample coverage used by coverage.

distances

A distance matrix or an object of class stats::dist

gamma

If TRUE, \(\gamma\) diversity, i.e. diversity of the metacommunity, is computed.

normalize

If TRUE, phylogenetic is normalized: the height of the tree is set to 1.

rate

The decay rate of the exponential similarity.

similarities

A similarity matrix, that can be obtained by fun_similarity. Its default value is the identity matrix.

species_distribution

An object of class species_distribution.

tree

An ultrametric, phylogenetic tree. May be an object of class phylo_divent, ape::phylo, ade4::phylog or stats::hclust.

weights

The weights of the sites of the species distributions.

Value

Returns TRUE or stops if a problem is detected.

Details

The function compares the arguments passed to its parent function to the type they should be and performs some extra tests, e.g. probabilities must be positive and sum to 1. It stops if an argument is not correct.

The function is always called without arguments. Its arguments exist only for documentation.