check_divent_args

Checks the arguments of a function of the package divent

Usage

check_divent_args(
  abundances = NULL,
  alpha = NULL,
  as_numeric = NULL,
  check_arguments = NULL,
  correction = NULL,
  coverage_estimator = NULL,
  distances = NULL,
  estimator = NULL,
  gamma = NULL,
  global = NULL,
  jack_alpha = NULL,
  jack_max = NULL,
  k = NULL,
  level = NULL,
  n_simulations = NULL,
  normalize = NULL,
  probability_estimator = NULL,
  q = NULL,
  r = NULL,
  rate = NULL,
  richness_estimator = NULL,
  sample_coverage = NULL,
  show_progress = NULL,
  similarities = NULL,
  species_distribution = NULL,
  tree = NULL,
  use.names = NULL,
  unveiling = NULL,
  weights = NULL,
  X = NULL
)

Arguments

abundances

an object of class abundances.

alpha

the risk level, 5% by default.

as_numeric

if TRUE, a number or a numeric vector is returned rather than a tibble.

check_arguments

if TRUE, the function arguments are verified. Should be set to FALSE to save time when the arguments have been checked elsewhere.

correction

the edge-effect correction to apply when estimating the number of neighbors or the K function with spatstat.explore::Kest. Default is "isotropic".

coverage_estimator

an estimator of sample coverage used by coverage.

distances

a distance matrix or an object of class stats::dist

estimator

an estimator of asymptotic entropy, diversity or richness.

gamma

if TRUE, \(\gamma\) diversity, i.e. diversity of the metacommunity, is computed.

global

if TRUE, a global envelope sensu Duranton2005divent is calculated.

jack_alpha

the risk level, 5% by default, used to optimize the jackknife order.

jack_max

the highest jackknife order allowed. Default is 10.

k

the order of Hurlbert's diversity.

level

the level of interpolation or extrapolation. It may be a sample size (an integer) or a sample coverage (a number between 0 and 1). If not NULL, the asymptotic estimator is ignored.

n_simulations

the number of simulations used to estimate the confidence envelope.

normalize

if TRUE, phylogenetic is normalized: the height of the tree is set to 1.

probability_estimator

a string containing one of the possible estimators of the probability distribution (see probabilities). Used only for extrapolation.

q

the order of diversity.

r

a vector of distances. If NULL accumulation is along n, else neighbors are accumulated in circles of radius r.

rate

the decay rate of the exponential similarity.

richness_estimator

an estimator of richness to evaluate the total number of species, see div_richness. used for interpolation and extrapolation.

sample_coverage

the sample coverage of x calculated elsewhere. Used to calculate the gamma diversity of meta-communities, see details.

show_progress

if TRUE, a progress bar is shown during long computations.

similarities

a similarity matrix, that can be obtained by fun_similarity. Its default value is the identity matrix.

species_distribution

an object of class species_distribution.

tree

an ultrametric, phylogenetic tree. May be an object of class phylo_divent, ape::phylo, ade4::phylog or stats::hclust.

use.names

if TRUE, the names of the species_distribution are kept in the matrix or vector they are converted to.

unveiling

a string containing one of the possible unveiling methods to estimate the probabilities of the unobserved species (see probabilities). Used only for extrapolation.

weights

the weights of the sites of the species distributions.

X

a spatialized community (A dbmss::wmppp object with PointType values as species names.)

Value

Returns TRUE or stops if a problem is detected.

Details

The function compares the arguments passed to its parent function to the type they should be and performs some extra tests, e.g. probabilities must be positive and sum to 1. It stops if an argument is not correct.

The function is always called without arguments. Its arguments exist only for documentation.